session29/movies/requirements.txt
django
pymongo
djangorestframework